Cover generated image reference normalization
